python3/wallpaper/requirements.txt
Pillow==5.3.0
pywin32==224
